+++ cordova-app/platforms/android/CordovaLib/ant-gen/R.java.d
@@ -1,1 +1,0 @@
- : /home/matt/Programming/Workspace/MiniMobileDeviceLab/cordova-app/platforms/android/CordovaLib/ant-build/AndroidManifest.xml \