+++ cordova-app/platforms/android/ant-build/AndroidManifest.xml.d
@@ -1,2 +1,0 @@
-/home/matt/Programming/Workspace/MiniMobileDeviceLab/cordova-app/platforms/android/ant-build/AndroidManifest.xml : \
-/home/matt/Programming/Workspace/MiniMobileDeviceLab/cordova-app/platforms/android/AndroidManifest.xml \