+++ cordova-app/platforms/android/ant-build/DeviceLab-debug-unaligned.apk.d
@@ -1,5 +1,0 @@
-/home/matt/Programming/Workspace/MiniMobileDeviceLab/cordova-app/platforms/android/ant-build/DeviceLab-debug-unaligned.apk : \
-/home/matt/Programming/Workspace/MiniMobileDeviceLab/cordova-app/platforms/android/ant-build/DeviceLab.ap_ \
-/home/matt/Programming/Workspace/MiniMobileDeviceLab/cordova-app/platforms/android/ant-build/classes.dex \
-/home/matt/Programming/Workspace/MiniMobileDeviceLab/cordova-app/platforms/android/CordovaLib/ant-build/classes.jar \
-/home/matt/Programming/Workspace/MiniMobileDeviceLab/cordova-app/platforms/android/libs/google-play-services.jar \