+++ cordova-app/platforms/android/ant-gen/co/uk/gauntface/mobile/devicelab/BuildConfig.java
@@ -1,6 +1,0 @@
-/** Automatically generated file. DO NOT MODIFY */
-package co.uk.gauntface.mobile.devicelab;
-
-public final class BuildConfig {
-    public final static boolean DEBUG = true;
-}