+++ cordova-app/plugins/org.apache.cordova.device/src/windows8/DeviceProxy.js
@@ -40,11 +40,11 @@ module.exports = {
         }
 
         setTimeout(function () {
-            win({ platform: "windows8", version: "8", uuid: deviceId, cordova: '0.0.0', model: window.clientInformation.platform });
+            win({ platform: "windows8", version: "8", uuid: deviceId, model: window.clientInformation.platform });
         }, 0);
     }
 
 };
 
-require("cordova/windows8/commandProxy").add("Device", module.exports);
+require("cordova/exec/proxy").add("Device", module.exports);