+++ node/config.js
@@ -1,4 +1,0 @@
-exports.dbHost = 'localhost';
-exports.dbUsername = 'root';
-exports.dbPassword = 'root';
-exports.dbPort = 8891;